update README daemon compile instructions
Also added a TODO list that needs to be addressed as there is some critical information that is outdated in regards the implemented technology and features that reside now in Espers. This is in regards especially to the VRX and Velocity systems including not having an xnode section at all.